# task-manager/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get a list of tasks with optional filtering by status",
                "operationId": "471d938a3bfaeb6fbd981d7c3e2728b0",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter tasks by their status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "TODO",
                                "IN_PROGRESS",
                                "COMPLETED"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of tasks"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Create a new task",
                "operationId": "f803ebdb2300acdb7e93bbe8d0b9a2e8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "status",
                                    "importance",
                                    "deadline"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "New Task"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "This is a description for a new task."
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "TODO",
                                            "IN_PROGRESS",
                                            "COMPLETED"
                                        ],
                                        "example": "TODO"
                                    },
                                    "importance": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1,
                                        "example": 3
                                    },
                                    "deadline": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-12-01 12:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task created successfully"
                    },
                    "400": {
                        "description": "Invalid data provided",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "importance": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Укажите важность задачи."
                                                    }
                                                },
                                                "title": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Заголовок задачи обязателен."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tasks/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get a single task by ID",
                "operationId": "c1a49a82b2ff84fa873fbf11566d29b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the task to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task found"
                    },
                    "404": {
                        "description": "Task not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Update an existing task",
                "operationId": "d09a073c4ac85fb554bcdf51e086e73a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the task to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "status",
                                    "importance",
                                    "deadline"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Task Title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated task description."
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "TODO",
                                            "IN_PROGRESS",
                                            "COMPLETED"
                                        ],
                                        "example": "IN_PROGRESS"
                                    },
                                    "importance": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1,
                                        "example": 4
                                    },
                                    "deadline": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-12-01T12:00:00Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Task updated successfully"
                    },
                    "404": {
                        "description": "Task not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Delete a task",
                "operationId": "01d7d04febb32ea2ecc7222450013dd5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the task to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Task deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Task not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tasks/priority": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get prioritized list of tasks",
                "operationId": "ee86301b1fbea7c7db28237e63a594a6",
                "responses": {
                    "200": {
                        "description": "List of tasks sorted by priority"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Tasks",
            "description": "Operations related to tasks"
        }
    ]
}